Add search for category
